`aframe serve`: rewrite (WIP)
