+++ commands/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   'deploy': require('./deploy.js'),
-  'serve': require('./serve.js'),
   'submit': require('./submit.js')
 };
