--- commands/serve.js
@@ -1,1 +1,0 @@
-module.exports = require('../lib/serve.js').serve;
